Adds new Events / Event UI, Event API, etc.
Adds new Events and Event UI pages. They are not complete but they are a
good start. The Event page is currently missing the participants widget, I
am still working on it. This commit includes the participants hub and the
participants api controller which will be used to build the new
participants widget. I also added momentjs to prettify times.